Use DEB/RPM Packages instead of compiling from source (#2)
We want to use deb/rpm packages because it's easier and they are amazingly faster to deploy. compiling python, alone, takes forever, and is unnecessary to do on every machine. We can always include another package, while it will increase the size of the distribution, since they're not all being unpacked on every machine, but rather the waste is being deleted, it's fine.

This adds that, and the mechanisms to release it and build it.